♻️[Refactor] FiterSection 유틸 함수 분리 #204
