+++ src/components/common-layout/FilterSection.tsx
@@ -5,6 +5,7 @@ import FilterCheckbox from '@/components/filter-checkbox/FilterCheckbox';
 import { ChangeEvent, Dispatch, SetStateAction, useState } from 'react';
 import SortingButton from '@/components/sorting-button/SortingButton';
 import { BookClub, BookClubParams } from '../../types/bookclubs';
+import { getMeetingType, getMemberLimit } from '@/lib/utils/filterUtils';
 
 interface CategoryTabsProps {
   bookClubs: BookClub[];
@@ -19,10 +20,9 @@ function FilterSection({
 }: CategoryTabsProps) {
   const [showAvailableOnly, setShowAvailableOnly] = useState(false); // 신청가능
 
-  const filterCheckboxHandler = (e: ChangeEvent<HTMLInputElement>) => {
+  const toggleAvailableOnly = (e: ChangeEvent<HTMLInputElement>) => {
     const isChecked = e.target.checked;
     setShowAvailableOnly(isChecked);
-    // onFilterChange(showAvailableOnly); // 백엔드에서 처리하는 로직 필요?
     if (isChecked) {
       const filteredBookClubs = bookClubs.filter(
         (club) => club.memberCount < club.memberLimit,
@@ -32,48 +32,16 @@ function FilterSection({
   };
 
   const handleMemberLimitChange = (selectedValue: string | undefined) => {
-    let memberLimit = undefined;
-
-    switch (selectedValue) {
-      case 'TWO_FOUR':
-        memberLimit = 4;
-        break;
-      case 'FIVE_SEVEN':
-        memberLimit = 7;
-        break;
-      case 'EIGHT_TEN':
-        memberLimit = 10;
-        break;
-      case 'OVER_ELEVEN':
-        memberLimit = 11;
-        break;
-      default:
-        memberLimit = undefined;
-    }
-
-    onFilterChange({ memberLimit: memberLimit });
+    const memberLimit = getMemberLimit(selectedValue);
+    onFilterChange({ memberLimit });
   };
 
-  const onChangeOnOffSelection = (selectedLabel: string | undefined) => {
-    const validValues: BookClubParams['meetingType'][] = [
-      'ALL',
-      'ONLINE',
-      'OFFLINE',
-    ];
-
-    if (
-      selectedLabel &&
-      validValues.includes(selectedLabel as BookClubParams['meetingType'])
-    ) {
-      onFilterChange({
-        meetingType: selectedLabel as BookClubParams['meetingType'],
-      });
-    } else {
-      onFilterChange({ meetingType: undefined });
-    }
+  const handleMeetingTypeChange = (selectedLabel: string | undefined) => {
+    const meetingType = getMeetingType(selectedLabel);
+    onFilterChange({ meetingType });
   };
 
-  const onChangeOrderSelection = (order: string) => {
+  const setSortingOrder = (order: string) => {
     onFilterChange({ order: order as BookClubParams['order'] });
   };
 
@@ -83,7 +51,7 @@ function FilterSection({
       <div className="flex items-center gap-x-2">
         <DropDown
           variant="onOff"
-          onChangeSelection={onChangeOnOffSelection}
+          onChangeSelection={handleMeetingTypeChange}
           aria-label="온라인/오프라인 선택"
         />
         <DropDown
@@ -94,13 +62,13 @@ function FilterSection({
         <FilterCheckbox
           label="신청가능"
           checked={showAvailableOnly}
-          onChange={filterCheckboxHandler}
+          onChange={toggleAvailableOnly}
           aria-label="신청가능 필터"
         />
       </div>
       <SortingButton
         variant="byDeadline"
-        onClickSorting={onChangeOrderSelection}
+        onClickSorting={setSortingOrder}
         aria-label="마감임박 정렬"
       />
     </div>
